requirements.txt
asgiref==3.7.2
autopep8==2.0.2
Django==4.2.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
Pillow==10.0.0
pycodestyle==2.11.0
pytz==2023.3
sqlparse==0.4.4
